refactor(util): move class
